style(speech): split SpeechAnalyzer live session
